--- app/.env
@@ -1,2 +1,0 @@
-#Inserir aqui a chave da OPENAI
-OPENAI_API_KEY=
